backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.8",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.7",
    "markdown>=3.10.1",
    "markdown-pdf>=1.11",
    "mermaidian>=0.1.2",
    "python-dotenv>=1.2.1",
]
